move structs to memtable
